Use the documentation template for documentation build and publish
