+++ screwdriver.yaml
@@ -32,7 +32,7 @@ jobs:
             - postupdate_version: |
                 PACKAGE_VERSION="`meta get package.version`"
                 echo $PACKAGE_VERSION
-            - postvalidate_code: meta set package.version "$PACKAGE_VERSION"
+            - preend: meta set package.version "$PACKAGE_VERSION"
         requires: [~commit, ~pr]
 
     validate_type:
@@ -47,14 +47,14 @@ jobs:
 
     generate_version:
         template: python/generate_version
-        requires: [~commit, ~pr]
+        requires: [validate_test, validate_lint, validate_codestyle, validate_dependencies, validate_security, validate_type]
 
     publish_test_pypi:
         template: python/package_python
         environment:
             PUBLISH: True
             TWINE_REPOSITORY_URL: https://test.pypi.org/legacy/
-        requires: [generate_version, validate_test, validate_lint, validate_codestyle, validate_dependencies, validate_security, validate_type]
+        requires: [generate_version]
 
     verify_test_package:
         template: python/validate_pypi_package
@@ -64,7 +64,7 @@ jobs:
             - update_version: |
                 PACKAGE_VERSION="`meta get package.version`"
                 echo $PACKAGE_VERSION
-            # - postvalidate_code: meta set package.version "$PACKAGE_VERSION"
+            - preend: meta set package.version "$PACKAGE_VERSION"
         requires: [publish_test_pypi]
 
     publish_documentation:
